Move AscendTrigger into OtherEntersBattlefieldTrigger.java
Creating a new instance of AscendTrigger would hog memory.
